src/graphql-client-python/requirements.txt
click
gql
requests
aiohttp
websockets
PyJWT
